Add derivative features, slight refactor, and update data exploration nb to reflect
