+++ soln/data_loader.py
@@ -1,26 +1,23 @@
 import pandas as pd
-import numpy as np
+
+from soln.features import operate_on_copy
+from soln.features import add_space_derivatives, add_force_magnitudes
 
 
-# Some useful decorators
-def operate_on_copy(fn):
+def load(csv_fname, raw=False):
     """
-    Any data manipulation should be done on a fresh copy of the dataframe.
-    """
-    def inner(df, *args, **kwargs):
-        df_copy = df.copy()
-        return fn(df_copy, *args, **kwargs)
-
-    return inner
-
-
-@operate_on_copy
-def clean_df(initial_df):
-    """
-    Clean the input dataframe.
+    Load a csv data file, clean it, and add all features.
     """
 
-    return initial_df.copy()
+    df = pd.read_csv(csv_fname)
+
+    if raw:
+        return df
+
+    return df.pipe(trim_start)\
+             .pipe(trim_end)\
+             .pipe(add_force_magnitudes)\
+             .pipe(add_space_derivatives)
 
 
 @operate_on_copy
@@ -61,37 +58,3 @@ def trim_end(df, force_threshold=50):
             return df.iloc[:-i]
 
     assert False, "trim_end couldn't find the end of the build!"
-
-
-@operate_on_copy
-def add_force_magnitudes(df):
-    """
-    Add a column for |<fx, fy, fz>| (both R1 and R2).
-    """
-
-    for idx in [1, 2]:
-        x = df[f"fx_{idx}"]
-        y = df[f"fy_{idx}"]
-        z = df[f"fz_{idx}"]
-
-        df[f"f{idx}_magnitude"] = (x*x + y*y + z*z).map(np.sqrt)
-
-    return df
-
-
-def load_data(csv_fname, clean=False):
-    """
-    Load in csv data to a pandas dataframe, clean it, and (optionally)
-    add any enrichments that shouldn't be part of the modeling pipeline.
-    """
-
-    df = pd.read_csv(csv_fname)
-
-    if not clean:
-        return df
-
-    cleaned = df.pipe(trim_start)\
-                .pipe(trim_end)\
-                .pipe(add_force_magnitudes)
-
-    return cleaned
