+++ soln/features.py
@@ -1,0 +1,46 @@
+import numpy as np
+
+
+def operate_on_copy(fn):
+    """
+    Any data manipulation should be done on a fresh copy of the dataframe.
+    """
+
+    def inner(df, *args, **kwargs):
+        df_copy = df.copy()
+        return fn(df_copy, *args, **kwargs)
+
+    return inner
+
+
+@operate_on_copy
+def add_space_derivatives(df, fields=['x_enc_1', 'y_enc_1', 'z_enc_1']):
+    """
+    Compute derivatives and add as a new column
+    in the dataframe (col name simply prepended with 'd').
+    """
+
+    def avg_gradient(x):
+        i0, i1 = x.index[0], x.index[1]
+        return (x[i1] - x[i0]) / (len(x) - 1.0)
+
+    for field in fields:
+        df['d' + field] = df[field].rolling(2).apply(avg_gradient)
+
+    return df
+
+
+@operate_on_copy
+def add_force_magnitudes(df):
+    """
+    Add a column for |<fx, fy, fz>| (both R1 and R2).
+    """
+
+    for idx in [1, 2]:
+        x = df[f"fx_{idx}"]
+        y = df[f"fy_{idx}"]
+        z = df[f"fz_{idx}"]
+
+        df[f"f{idx}_magnitude"] = (x*x + y*y + z*z).map(np.sqrt)
+
+    return df
